requirements.txt
Django==5.2
djangorestframework==3.15.0
pytest==7.4.3
pytest-django==4.7.0